refactor: simplify garbage collection implementation (without mtime)

+++ crates/tabby-scheduler/src/code/repository.rs
@@ -1,4 +1,5 @@
 use std::{
+    collections::HashSet,
     fs::{self, OpenOptions},
     path::{Path, PathBuf},
     process::Command,
@@ -42,38 +43,7 @@ impl RepositoryExt for RepositoryConfig {
                 }
             }
         }
-
-        if finished {
-            debug!("Repository `{}` is up to date", dir.display());
-            touch(&get_last_repository_sync_filepath(&self.dir()));
-        }
     }
-}
-
-fn touch(path: &Path) {
-    std::fs::create_dir_all(path.parent().expect("Failed to read parent"))
-        .expect("Failed to create directory");
-    OpenOptions::new()
-        .write(true)
-        .create(true)
-        .open(path)
-        .expect("Failed to touch file");
-}
-
-fn get_last_repository_sync_filepath(path: &Path) -> PathBuf {
-    path.join(".git").join("tabby").join("last_repository_sync")
-}
-
-fn get_last_repository_sync_time(path: &Path) -> std::time::SystemTime {
-    let filepath = get_last_repository_sync_filepath(path);
-    if !filepath.exists() {
-        touch(&filepath);
-    }
-
-    fs::metadata(&filepath)
-        .expect("Failed to read metadata")
-        .modified()
-        .expect("Failed to read modified")
 }
 
 fn pull_remote(path: &Path) -> bool {
@@ -105,11 +75,11 @@ pub fn sync_repository(repository: &RepositoryConfig) {
     } else {
         repository.sync();
     }
-
-    garbage_collection();
 }
 
-fn garbage_collection() {
+pub fn garbage_collection(repositories: &[RepositoryConfig]) {
+    let names = repositories.iter().map(|r| r.dir()).collect::<HashSet<_>>();
+
     for file in fs::read_dir(repositories_dir())
         .expect("Failed to read repository dir")
         .filter_map(Result::ok)
@@ -121,18 +91,11 @@ fn garbage_collection() {
             // There shouldn't be any files under repositories dir.
             fs::remove_file(file.path())
                 .unwrap_or_else(|_| panic!("Failed to remove file {:?}", filename))
-        } else if metadata.is_dir() {
-            let mtime = get_last_repository_sync_time(&file.path());
-
-            // if stale for 2 day, consider it as garbage.
-            let is_garbage = mtime.elapsed().unwrap_or_default().as_secs() > 2 * 24 * 60 * 60;
-
-            if is_garbage {
-                warn!("An unrelated directory {:?} was found in repositories directory, It will now be removed...", file.path().display());
-                fs::remove_dir_all(file.path()).unwrap_or_else(|_| {
-                    panic!("Failed to remove directory {:?}", file.path().display())
-                });
-            }
+        } else if metadata.is_dir() && !names.contains(&file.path()) {
+            warn!("An unrelated directory {:?} was found in repositories directory, It will now be removed...", file.path().display());
+            fs::remove_dir_all(file.path()).unwrap_or_else(|_| {
+                panic!("Failed to remove directory {:?}", file.path().display())
+            });
         }
     }
 }
